# api-docs-PRUEBA.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/eventos/{id}":{"get":{"tags":["evento-controller"],"operationId":"getEventoById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Evento"}}}}}},"put":{"tags":["evento-controller"],"operationId":"updateEvento","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventoRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Evento"}}}}}},"delete":{"tags":["evento-controller"],"operationId":"deleteEvento","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Evento"}}}}}}}},"/api/eventos":{"get":{"tags":["evento-controller"],"operationId":"getAllEventos","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Evento"}}}}}}},"post":{"tags":["evento-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventoRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EventoResponse"}}}}}}}},"components":{"schemas":{"EventoRequest":{"required":["fechaevento","lugar","nombre","participanteid","tipoevento"],"type":"object","properties":{"eventoId":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"tipoevento":{"type":"string"},"participanteid":{"type":"integer","format":"int64"},"fechaevento":{"type":"string","format":"date"},"lugar":{"type":"string"}}},"Evento":{"type":"object","properties":{"eventoid":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"tipoevento":{"type":"string"},"participanteid":{"type":"integer","format":"int64"},"fechaevento":{"type":"string","format":"date"},"lugar":{"type":"string"}}},"EventoResponse":{"type":"object","properties":{"eventoid":{"type":"integer","format":"int64"},"nombre":{"type":"string"}}}}}}